AI_backend/api/requirements.txt
fastapi
uvicorn
tensorflow
numpy
psycopg2-binary
prometheus_client
